+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/StaticDataLinkTypePacketFactory.java
@@ -9,8 +9,7 @@ package org.pcap4j.packet.factory;
 
 import org.pcap4j.packet.EthernetPacket;
 import org.pcap4j.packet.IllegalRawDataException;
-import org.pcap4j.packet.IpV4Packet;
-import org.pcap4j.packet.IpV6Packet;
+import org.pcap4j.packet.IpPacket;
 import org.pcap4j.packet.Packet;
 import org.pcap4j.packet.UnknownPacket;
 import org.pcap4j.packet.namednumber.DataLinkType;
@@ -46,16 +45,7 @@ extends AbstractStaticPacketFactory<DataLinkType> {
         public Packet newInstance(
           byte[] rawData, int offset, int length
         ) throws IllegalRawDataException {
-          if (length > 0) {
-            int ipVersion = (rawData[0] >> 4) & 0x0f;
-            if (ipVersion == 4) {
-              return IpV4Packet.newPacket(rawData, offset, length);
-            }
-            if (ipVersion == 6) {
-              return IpV6Packet.newPacket(rawData, offset, length);
-            }
-          }
-          return UnknownPacket.newPacket(rawData, offset, length);
+          return IpPacket.newPacket(rawData, offset, length);
         }
         @Override
         public Class<UnknownPacket> getTargetClass() {